refactor: use blueprint server throughout (#495)
