+++ src/http/request_context.rs
@@ -4,7 +4,8 @@ use derive_setters::Setters;
 use hyper::HeaderMap;
 
 use super::{DefaultHttpClient, Response, ServerContext};
-use crate::config::Server;
+use crate::blueprint::Server;
+use crate::config;
 
 #[derive(Setters)]
 pub struct RequestContext {
@@ -16,7 +17,10 @@ pub struct RequestContext {
 
 impl Default for RequestContext {
   fn default() -> Self {
-    RequestContext::new(DefaultHttpClient::default(), Server::default())
+    let config = config::Server::default();
+    //TODO: default is used only in tests. Drop default and move it to test.
+    let server = Server::try_from(config).unwrap();
+    RequestContext::new(DefaultHttpClient::default(), server)
   }
 }
 
@@ -52,7 +56,7 @@ impl RequestContext {
 impl From<&ServerContext> for RequestContext {
   fn from(server_ctx: &ServerContext) -> Self {
     let http_client = server_ctx.http_client.clone();
-    Self::new(http_client, server_ctx.server.clone())
+    Self::new(http_client, server_ctx.blueprint.server.clone())
   }
 }
