+++ src/http/server.rs
@@ -20,22 +20,24 @@ fn graphiql() -> Result<Response<Body>> {
 }
 
 async fn graphql_request(req: Request<Body>, server_ctx: &ServerContext) -> Result<Response<Body>> {
-  let server = server_ctx.server.clone();
-  let allowed = server.upstream.allowed_headers.unwrap_or_default();
+  let server = server_ctx.blueprint.server.clone();
+  let allowed = server.upstream.get_allowed_headers();
   let headers = create_allowed_headers(req.headers(), &allowed);
   let bytes = hyper::body::to_bytes(req.into_body()).await?;
   let request: async_graphql_hyper::GraphQLRequest = serde_json::from_slice(&bytes)?;
   let req_ctx = Arc::new(RequestContext::from(server_ctx).req_headers(headers));
   let mut response = request.data(req_ctx.clone()).execute(&server_ctx.schema).await;
 
-  if server_ctx.server.enable_cache_control() {
+  if server_ctx.blueprint.server.enable_cache_control_header {
     if let Some(ttl) = req_ctx.get_min_max_age() {
       response = response.set_cache_control(ttl as i32);
     }
   }
   let mut resp = response.to_response()?;
-  if !server_ctx.custom_response_header.is_empty() {
-    resp.headers_mut().extend(server_ctx.custom_response_header.clone());
+  if !server_ctx.blueprint.server.response_headers.is_empty() {
+    resp
+      .headers_mut()
+      .extend(server_ctx.blueprint.server.response_headers.clone());
   }
 
   Ok(resp)
@@ -45,7 +47,16 @@ fn not_found() -> Result<Response<Body>> {
 }
 async fn handle_request(req: Request<Body>, state: Arc<ServerContext>) -> Result<Response<Body>> {
   match *req.method() {
-    hyper::Method::GET if state.server.enable_graphiql.as_ref() == Some(&req.uri().path().to_string()) => graphiql(),
+    hyper::Method::GET
+      if state
+        .blueprint
+        .server
+        .enable_graphiql
+        .as_ref()
+        .map_or(false, |s| s.as_str() == req.uri().path()) =>
+    {
+      graphiql()
+    }
     hyper::Method::POST if req.uri().path() == "/graphql" => graphql_request(req, state.as_ref()).await,
     _ => not_found(),
   }
@@ -61,9 +72,8 @@ fn create_allowed_headers(headers: &HeaderMap, allowed: &BTreeSet<String>) -> He
   new_headers
 }
 pub async fn start_server(config: Config) -> Result<()> {
-  let server = config.server.clone();
   let blueprint = Blueprint::try_from(&config).map_err(CLIError::from)?;
-  let state = Arc::new(ServerContext::new(blueprint.clone(), server));
+  let state = Arc::new(ServerContext::new(blueprint.clone()));
   let make_svc = make_service_fn(move |_conn| {
     let state = Arc::clone(&state);
     async move { Ok::<_, anyhow::Error>(service_fn(move |req| handle_request(req, state.clone()))) }
